requirements.txt
uvicorn
httpx
requests
Pillow
fastapi
openai